Move export related functions to export_snapshot.py
